moves wms 1.3.0 xml_mapper correct naming file (test files also)
