+++ backend/tests/django/settings.py
@@ -1,0 +1,3 @@
+import os
+
+DJANGO_TEST_ROOT_DIR = os.path.dirname(os.path.abspath(__file__))
